投票系統/requirements.txt
Flask
opencv-python
pytesseract
pybase64
pycryptodome
pathlib